refactor: Reorganised files to improve file separation
