+++ matchingService/src/QueueService/Producer.ts
@@ -1,6 +1,6 @@
 import { Channel, Replies } from "amqplib";
-import MatchRequest from "./MatchRequest";
-import { MessageHeader } from "./MessageHeaders";
+import MatchRequest from "../models/MatchRequest";
+import { MessageHeader } from "../models/MessageHeaders";
 import { v4 as uuidv4 } from 'uuid';
 
 var responseQueue: string = "responseQueue";
